+++ shell_scripts/deploy/go_rtmp1.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+#Author: chenchanghui@lizi-inc.com
+#Date: 2016.05.31
+#Desc:
+#       SSH LOGIN 101.201.100.208 
+ssh root@101.201.100.208