+++ shell_scripts/changelog.md
@@ -1,6 +1,12 @@
 # CHANGE LOG
 
 
+## 2019.06.10
++ deploy: 部署工具集
++ net.sh: 网卡流量监测工具
++ access_mysql.sh: 访问MySQL数据库
++ create_ssl_ca.sh: 自签发CA根证书，用于开发或测试，需要安装openssl
+
 ## 2019.05.30
-+ format_date_as_string.sh, 格式化时间
-+ get_file_name_without_suffix_in_path.sh, 在执行ls命令后获取不带后缀名的文件名称
++ format_date_as_string.sh: 格式化时间
++ get_file_name_without_suffix_in_path.sh: 在执行ls命令后获取不带后缀名的文件名称